+++ extendr-api/src/robj/mod.rs
@@ -799,7 +799,10 @@ make_typed_slice!(c64, COMPLEX, CPLXSXP);
 make_typed_slice!(Rcplx, COMPLEX, CPLXSXP);
 make_typed_slice!(Rcomplex, COMPLEX, CPLXSXP);
 
-/// These are helper functions which give access to common properties of R objects.
+/// Provides access to the attributes of an R object.
+///
+/// The `Attribute` trait provides a consistent interface to getting, setting, and checking for the presence of attributes in an R object.
+///
 #[allow(non_snake_case)]
 pub trait Attributes: Types + Length {
     /// Get a specific attribute as a borrowed `Robj` if it exists.
@@ -852,11 +855,12 @@ pub trait Attributes: Types + Length {
     /// ```
     /// use extendr_api::prelude::*;
     /// test! {
-    ///    let mut robj = r!("hello").set_attrib(sym!(xyz), 1)?;
+    ///    let mut robj = r!("hello");
+    ///    robj.set_attrib(sym!(xyz), 1)?;
     ///    assert_eq!(robj.get_attrib(sym!(xyz)), Some(r!(1)));
     /// }
     /// ```
-    fn set_attrib<N, V>(&mut self, name: N, value: V) -> Result<Robj>
+    fn set_attrib<N, V>(&mut self, name: N, value: V) -> Result<&mut Self>
     where
         N: Into<Robj>,
         V: Into<Robj>,
@@ -865,17 +869,14 @@ pub trait Attributes: Types + Length {
         let value = value.into();
         unsafe {
             let sexp = self.get_mut();
-            single_threaded(|| {
-                catch_r_error(|| Rf_setAttrib(sexp, name.get(), value.get()))
-                    // FIXME: there is no reason to re-wrap this, as this mutates
-                    // the input `self`, and returns another pointer to the same
-                    // object
-                    .map(|_| Robj::from_sexp(sexp))
-            })
+            let result =
+                single_threaded(|| catch_r_error(|| Rf_setAttrib(sexp, name.get(), value.get())));
+            let result = result.map(|_| self);
+            result
         }
     }
 
-    /// Get the names attribute as a string iterator if one exists.
+    /// Get the `names` attribute as a string iterator if one exists.
     /// ```
     /// use extendr_api::prelude::*;
     /// test! {
@@ -892,12 +893,12 @@ pub trait Attributes: Types + Length {
         }
     }
 
-    /// Return true if this object has names.
+    /// Return true if this object has an attribute called `names`.
     fn has_names(&self) -> bool {
         self.has_attrib(wrapper::symbol::names_symbol())
     }
 
-    /// Set the names attribute from a string iterator.
+    /// Set the `names` attribute from a string iterator.
     ///
     /// Returns `Error::NamesLengthMismatch` if the length of the names does
     /// not match the length of the object.
@@ -905,12 +906,13 @@ pub trait Attributes: Types + Length {
     /// ```
     /// use extendr_api::prelude::*;
     /// test! {
-    ///     let mut obj = r!([1, 2, 3]).set_names(&["a", "b", "c"]).unwrap();
+    ///     let mut obj = r!([1, 2, 3]);
+    ///     obj.set_names(&["a", "b", "c"]).unwrap();
     ///     assert_eq!(obj.names().unwrap().collect::<Vec<_>>(), vec!["a", "b", "c"]);
     ///     assert_eq!(r!([1, 2, 3]).set_names(&["a", "b"]), Err(Error::NamesLengthMismatch(r!(["a", "b"]))));
     /// }
     /// ```
-    fn set_names<T>(&mut self, names: T) -> Result<Robj>
+    fn set_names<T>(&mut self, names: T) -> Result<&mut Self>
     where
         T: IntoIterator,
         T::IntoIter: ExactSizeIterator,
@@ -927,7 +929,7 @@ pub trait Attributes: Types + Length {
         }
     }
 
-    /// Get the dim attribute as an integer iterator if one exists.
+    /// Get the `dim` attribute as an integer iterator if one exists.
     /// ```
     /// use extendr_api::prelude::*;
     /// test! {
@@ -945,7 +947,7 @@ pub trait Attributes: Types + Length {
         }
     }
 
-    /// Get the dimnames attribute as a list iterator if one exists.
+    /// Get the `dimnames` attribute as a list iterator if one exists.
     /// ```
     /// use extendr_api::prelude::*;
     /// test! {
@@ -962,7 +964,7 @@ pub trait Attributes: Types + Length {
         }
     }
 
-    /// Get the class attribute as a string iterator if one exists.
+    /// Get the `class` attribute as a string iterator if one exists.
     /// ```
     /// use extendr_api::prelude::*;
     /// test! {
@@ -979,19 +981,20 @@ pub trait Attributes: Types + Length {
         }
     }
 
-    /// Set the class attribute from a string iterator, and returns the same
+    /// Set the `class` attribute from a string iterator, and return the same
     /// object.
     ///
     /// May return an error for some class names.
     /// ```
     /// use extendr_api::prelude::*;
     /// test! {
-    ///     let mut obj = r!([1, 2, 3]).set_class(&["a", "b", "c"])?;
+    ///     let mut obj = r!([1, 2, 3]);
+    ///     obj.set_class(&["a", "b", "c"])?;
     ///     assert_eq!(obj.class().unwrap().collect::<Vec<_>>(), vec!["a", "b", "c"]);
     ///     assert_eq!(obj.inherits("a"), true);
     /// }
     /// ```
-    fn set_class<T>(&mut self, class: T) -> Result<Robj>
+    fn set_class<T>(&mut self, class: T) -> Result<&mut Self>
     where
         T: IntoIterator,
         T::IntoIter: ExactSizeIterator,
@@ -1018,7 +1021,7 @@ pub trait Attributes: Types + Length {
         }
     }
 
-    /// Get the levels attribute as a string iterator if one exists.
+    /// Get the `levels` attribute as a string iterator if one exists.
     /// ```
     /// use extendr_api::prelude::*;
     /// test! {